Make CPlayer a part of Kamino again
CPlayer is heavily integrated with Apollo Kamino. Any elements of portability initially built into CPlayer have now been removed or counteracted in lieu of haste and convenience.
There appears to be little public interest in using CPlayer as a player externally from ApolloTV so it makes sense to simplify things in this manner.

Also much of the work in the underlying 'core' of CPlayer (i.e. video in Flutter) is going to be going in to the flutter_vlc_plugin.
As for those interested specifically in the CPlayer UI, it should be easy enough to copy the CPlayer module over from the Kamino app and figure out the dependencies :P.